refactor: global directory 구조 수정
- global 디렉토리 구조 수정

#32
