+++ src/main/kotlin/com/bandage/v1/domain/band/repository/BandRepositoryCustom.kt
@@ -1,7 +1,7 @@
 package com.bandage.v1.domain.band.repository
 
 import com.bandage.v1.domain.band.model.Band
-import com.bandage.v1.global.common.dto.CursorResponse
+import com.bandage.v1.global.common.response.CursorResponse
 import java.util.UUID
 
 interface BandRepositoryCustom {
