+++ src/main/kotlin/com/bandage/v1/global/error/exception/Exception.kt
@@ -1,0 +1,7 @@
+package com.bandage.v1.global.error.exception
+
+import com.bandage.v1.global.error.errorcode.ErrorCode
+
+open class Exception(
+    val errorCode: ErrorCode,
+) : RuntimeException(errorCode.message)
